refactor(Band~) : 객체 기능 분리
