+++ wabi/src/main/kotlin/com/wap/wabi/band/controller/BandQueryController.kt
@@ -2,20 +2,14 @@ package com.wap.wabi.band.controller
 
 import com.wap.wabi.auth.admin.service.AdminService
 import com.wap.wabi.auth.jwt.JwtTokenProvider
-import com.wap.wabi.band.payload.request.BandCreateRequest
-import com.wap.wabi.band.payload.request.BandUpdateRequest
 import com.wap.wabi.band.payload.response.BandStudentsData
-import com.wap.wabi.band.service.BandService
+import com.wap.wabi.band.service.BandQueryService
 import com.wap.wabi.common.payload.response.Response
 import io.swagger.v3.oas.annotations.Operation
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.PutMapping
-import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestHeader
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
@@ -23,7 +17,7 @@ import org.springframework.web.bind.annotation.RestController
 @RestController
 @RequestMapping("/api/bands/")
 class BandQueryController(
-    private val bandService: BandService,
+    private val bandQueryService: BandQueryService,
     private val jwtTokenProvider: JwtTokenProvider,
     private val adminService: AdminService
 ) {
@@ -32,40 +26,8 @@ class BandQueryController(
         summary = "밴드에 속한 학생 명단 조회"
     )
     fun getBandStudents(@PathVariable bandId: Long): ResponseEntity<Response> {
-        val response = Response.ok(data = BandStudentsData(bandService.getBandStudents(bandId = bandId)))
+        val response = Response.ok(data = BandStudentsData(bandQueryService.getBandStudents(bandId = bandId)))
 
-        return ResponseEntity(response, HttpStatus.OK)
-    }
-
-    @PostMapping("create")
-    @Operation(
-        summary = "밴드 생성"
-    )
-    fun createBand(
-        @RequestHeader("Authorization") token: String,
-        @RequestBody request: BandCreateRequest
-    ): ResponseEntity<Response> {
-        val adminName = jwtTokenProvider.getAdminNameByToken(token.removePrefix("Bearer "))
-        val adminId = adminService.getAdminId(adminName = adminName)
-        bandService.createBand(adminId = adminId, bandCreateRequest = request)
-
-        val response = Response.ok(message = "success create band")
-        return ResponseEntity(response, HttpStatus.OK)
-    }
-
-    @DeleteMapping("{bandId}")
-    @Operation(
-        summary = "밴드 삭제"
-    )
-    fun deleteBand(
-        @RequestHeader("Authorization") token: String,
-        @PathVariable bandId: Long
-    ): ResponseEntity<Response> {
-        val adminName = jwtTokenProvider.getAdminNameByToken(token.removePrefix("Bearer "))
-        val adminId = adminService.getAdminId(adminName = adminName)
-        bandService.deleteBand(adminId = adminId, bandId = bandId)
-
-        val response = Response.ok(message = "success delete band")
         return ResponseEntity(response, HttpStatus.OK)
     }
 
@@ -76,24 +38,7 @@ class BandQueryController(
     fun getBands(@RequestHeader("Authorization") token: String): ResponseEntity<Response> {
         val adminName = jwtTokenProvider.getAdminNameByToken(token.removePrefix("Bearer "))
         val adminId = adminService.getAdminId(adminName = adminName)
-        val response = Response.ok(data = bandService.getBands(adminId = adminId))
-        return ResponseEntity(response, HttpStatus.OK)
-    }
-
-    @PutMapping("")
-    @Operation(
-        summary = "밴드 수정"
-    )
-    fun updateBand(
-        @RequestHeader("Authorization") token: String,
-        @RequestBody request: BandUpdateRequest
-    ): ResponseEntity<Response> {
-        val adminName = jwtTokenProvider.getAdminNameByToken(token.removePrefix("Bearer "))
-        val adminId = adminService.getAdminId(adminName = adminName)
-        bandService.updateBand(adminId = adminId, bandUpdateRequest = request)
-
-        val response = Response.ok(message = "success update band")
-
+        val response = Response.ok(data = bandQueryService.getBands(adminId = adminId))
         return ResponseEntity(response, HttpStatus.OK)
     }
 
@@ -103,7 +48,7 @@ class BandQueryController(
     )
     fun getBandDetail(@PathVariable bandId: Long): ResponseEntity<Response> {
 
-        val response = Response.ok(data = bandService.getBandDetail(bandId = bandId))
+        val response = Response.ok(data = bandQueryService.getBandDetail(bandId = bandId))
 
         return ResponseEntity(response, HttpStatus.OK)
     }
